requirements.txt
asyncio
sanic
tortoise-orm
aiosqlite
aioredis
pyjwt
jsonschema
requests
pytest-sanic